+++ core/src/elements/ons-tabbar/index.js
@@ -145,6 +145,22 @@ export default class TabbarElement extends BaseElement {
    */
 
   /**
+   * @event swipe
+   * @description
+   *   [en]Fires when the tabbar swipes.[/en]
+   *   [ja][/ja]
+   * @param {Object} event
+   *   [en]Event object.[/en]
+   *   [ja]イベントオブジェクト。[/ja]
+   * @param {Number} event.index
+   *   [en]Current index.[/en]
+   *   [ja]現在アクティブになっているons-tabのインデックスを返します。[/ja]
+   * @param {Object} event.options
+   *   [en]Animation options object.[/en]
+   *   [ja][/ja]
+   */
+
+  /**
    * @attribute animation
    * @type {String}
    * @default none
@@ -251,6 +267,10 @@ export default class TabbarElement extends BaseElement {
       this._tabbarBorder = null;
       this._tabsRect = null;
     }
+
+    if (this._onSwipe) {
+      this.removeEventListener('swipe', this._onSwipe);
+    }
   }
 
   _normalizeEvent(event) {
@@ -298,7 +318,7 @@ export default class TabbarElement extends BaseElement {
       }
     }
 
-    this._onSwipe && this._onSwipe(index, options);
+    util.triggerElementEvent(this, 'swipe', { index, options });
   }
 
   _onRefresh() {
@@ -537,6 +557,12 @@ export default class TabbarElement extends BaseElement {
     if (value && !(value instanceof Function)) {
       util.throw(`"onSwipe" must be a function`)
     }
+
+    if (this._onSwipe) {
+      document.removeEventListener('swipe', this._onSwipe);
+    }
+    document.addEventListener('swipe', value);
+
     this._onSwipe = value;
   }